+++ ReleaseNotes.md
@@ -178,3 +178,4 @@ Please change your configuration according to [App.xaml](https://github.com/spie
 ### v3.3.0 (upcoming release)
 #### Features
 * Added `XmlnsDefinition` `https://spiegelp.github.io/MaterialDesignExtensions/winfx/xaml`
+* Consider `ResizeMode` for the window caption buttons